Add group resource with role mappings operations
