+++ lib/keycloak-admin/resource/group_resource.rb
@@ -1,0 +1,7 @@
+module KeycloakAdmin
+  class GroupResource < BaseRoleContainingResource
+    def resources_name
+      "groups"
+    end
+  end
+end
